Put the hack for sending msgs to NickServ back
It turns out some servers don't allow identification with `PASS` so we
have to send `PRIVMSG` to `NickServ` in `auto_cmds`, but we don't want
new tabs for that. So, this hack needs to stay until we implement named
command arguments and add a argument to `/cmd` for not creating tabs
(#66).