Add `--config` flag to specify config file location
Add DIY command line arg parsing to deal with a possible `--config` flag and treat all other arguments as servers to connect to. Also add a `config_path` field to the `Tiny`  struct so that the `reload` command is able to reload the config file (was not necessary previously as there was only 1 possible config location).