Send CAP LS and CAP REQ after USER and NICK
Before this commit, in SASL auth, the original order of messages is:

  CAP REQ :sasl
  NICK ...
  USER ...

Which follows the IRC spec: https://ircv3.net/specs/extensions/sasl-3.1.html

But it is found that if the nick is in use (perhaps you have logged in
in another computer), then we cannot authenticate to mozilla irc server,
while freenode is OK.  It is found that xchat works this way:

  NICK ...
  USER ...
  CAP REQ :sasl

When this is adopted, it is found that mozilla is ok but freenode is not if the
nick is NOT in use.  After adding a CAP LS before NICK both works.