Implement termbox library in Rust
This ports termbox's "cell" abstraction to Rust, removing the last piece
of C code from the repo.

Input and event handling of the original termbox was removed long time
ago, and we've been removing bits and pieces from the original termbox
since then, leaving only the bits we need. The remaining pieces was
basically just the cell abstraction, which we implement in Rust with
this patch, getting some help from termion.

Performance is on par with the original C implementation: libtiny_tui's "bench"
example, when passed a 19M file, runs in 20.1s in C implementation, and 20.2s in
Rust implementation. C implementation writes 918,141,709 bytes to stdout, Rust
implementation write 915,387,921 bytes (-0.2%). The difference may either be
better cursor location caching in the Rust version or a bug. So far I'm unable
to find any rendering bugs, so I'm inclined to think the former.

Co-authored-by: Emilio González <emigr2k1@gmail.com>