Try both IPv4 and IPv6 addresses when domain resolves to both
See #144 for the problem this solves. When a domain name resolves to
IPv4 and IPv6 addresses we now try to connect to both, concurrently, and
continue with the one that connects first. This solves the problem of
user only being able to connect to IPv4 or IPv6 addresses.

An alternative would be to ask the user to choose whether to use IPv6 or
not, and default to IPv4. I think this version is better; it doesn't
require any input from the user (no need for a new tinyrc field) -- it
just works.

Fixes #144