Simplify UI task termination logic, fix a race between clients and UI
Quitting tiny (C-c in TUI) does not mean all clients will immediately
stop: they'll send a QUIT message to the server, and depending on
scheduling they may even receive a message or two. In those cases if we
return from TUI right away when quitting we'd get panics in client code
because client assumes the receiving end of the connection event channel
is always open until the client returns.

We now keep the UI task running until all client channels are dropped.
Because the UI task is the owner of the TUI this avoids the race (TUI is
awailable as long as there's at least one client running) by keeping the
TUI alive.

A good side effect of this is that after C-c we won't be waiting in the
shell for the process to exit (i.e. clients to send QUIT), instead we'll
see the TUI.