Changes to state to handle auto-joining channels that have +R.
If the server sends back a 477 reply, meaning that we can't join the
channel without being registered and identified with NickServ, then we
check to see if nickserv_ident is configured. If it is configured then
we wait for 10 seconds and retry joining the channel. We do this 3 times
and then stop.

Removed some unwraps. Changed UX to alert user of retries

Added new Event to libtiny_client to send a channel message when a 477
reply comes from the server and we want to tell the user that a retry
will be made.

Also, the 477 reply from the server will show up in a channel tab which
lets the user know what is going on instead of not opening a tab at all.

Closing a tab that is in the process of joining a channel kills join
task.
Only send PART message to channel when we are actually joined (in the
channel).

PR feedback changes. Removed optional Sender in JoinState. Cleaned up
log messages. Moved async lambda into own function.

Added doc comment to JoinState. Renamed Initial state to NotJoined

Fixed bug that delays quitting when join task is running

More PR feedback