Handle 477 responses to re-join channels
If the server sends back a 477 reply, meaning that we can't join the
channel without being registered and identified with NickServ, then we
check to see if nickserv_ident is configured. If it is configured then
we wait for 10 seconds and retry joining the channel. We do this 3 times
and then stop.


Fixes #236