Improve prefix parsing and ambiguous prefix handling (#249)
Previously when we see an ambiguous prefix (e.g. "localhost") we'd assume
server. We now have a prefix variant Ambiguous for such cases and resolving
the ambiguity is left to the caller.

When there's an ambiguity we try to do the reasonable thing depending on the
type of the message:

- When the message is a numeric reply we assume server.
- In all other cases (e.g. in JOIN/PART/QUIT/PRIVMSG) we assume nick.

Other changes:

- Removed wildcard patterns to make the code future-proof.
- Refactored some logging and comments.

Fixes #247