Fix blocking during connection initialization
We apply the same idea used in resolve_addr and wait to allow responding
to user commands like /connect while opening a socket to the server.

Fixes #255

The patch is larger than necessary as I also refactored the existing
code for resolve_addr and wait a little bit, to allow code reuse and
simplify things a little bit.

We should also add a timer to the connect task. That's left to another
commit.