Fix case (in)sensitivity of channel names
According to RFC 2812 channel names are case insensitive, with special rules.
Section 2.2 says:

> Because of IRC's Scandinavian origin, the characters {}|^ are considered to be
> the lower case equivalents of the characters []\~, respectively. This is a
> critical issue when determining the equivalence of two nicknames or channel
> names.

RFC doesn't specify an encoding for characters outside of ASCII so it's unclear
what to do for channels with non-ASCII characters. However it seems like most
servers don't compare non-ASCII characters case insensitively (for example, on
Unreal3.2.10.3-gs #ööö and #ÖÖÖ are different channels but #ooo and #OOO are the
same), so the rules are:

- Only ASCII characters are compared case insensitively
- Lowercase versions of '[', ']', '\', and '~' are '{', '}', '|', and '^',
  respectively.

See also #248 for discussion.

To implement this we introduce a new crate, libtiny_common, that implements two
types: ChanName and ChanNameRef. They're just newtypes for String and str,
respectively. Trait implementation for Eq and Hash work case insensitively
according to the rules above.

Other creates are refactored to use these types whevher we use a String/&str for
channel names. As a result channel name comparisons are fixed everywhere in the
code base simply by mechanically replacing the types (with some minor
refactoring in a few places).

I'm not quite happy with introducing a new crate just for this purpose. We may
want to move these types to libtiny_wire, but this will do for now.

Fixes #248