term_input: Enable non-blocking mode
Normally term_input is used with non-canonical input with VMIN=0 and
VTIME=0. On Linux in this mode stdin does not block (returns 0 when it's
not ready for reading, see [1] and [2]). However on WSL it doesn't work
that way and we have to set stdin to non-blocking mode to be able to do
`read(stdin)` without blocking.

We now set stdin to non-blocking mode when creating an `Input` and
restore the old flags when dropping it.

Fixes #269

[1]: https://www.gnu.org/software/libc/manual/html_node/Canonical-or-Not.html
[2]: https://www.gnu.org/software/libc/manual/html_node/Noncanonical-Input.html