Message alignment on draw
Instead of applying padding to the actual data in the line buffers, this
applies the correct padding on the fly during line height calculation
and drawing.

Not a very good solution since I did this in order to be able to switch
layouts on the fly (#321) but it still doesn't work because we don't always
align every line (i.e topic line, server stuff).