Added /join cmd parameters
When joining a channel using the /join command, the tab is created
before the server is actually joined so we can effectively set the
channel config (ex. ignore the tab, set notifier).

Due to this, I had to add handling for REPLY messages that occur when a
channel cannot be joined but the tab has already been created. Now an
error message shows on the chan tab, giving the reason why the chan
can't be joined.