Add new tab type aka MsgSource/Target
Added a new "Misc" tab type that will genericize "mentions" and
"help" tab, and what ever else we want to add. This type of tab gets
created to the right of the current tab and came move freely around all
tabs. Moving a server tab captures a misc tab in its group.