Move cargo release configuration to CI
By setting these in Cargo.toml everyone is forced to build with them. It
is better to leave the defaults to Cargo and the user building the
project. If there are important gains in binary size or speed  with
these options they can continue to be set for the official builds as
performed by CI.